Switch to `f32` for `GeometryBuffer` depth (#381)
This makes it easier to send the `GeometryBuffer` to the GPU as a
`f32x4` texture. I previously just used bitcasting, but It turns out
that a small `u32` produces a denormalized float, and some GPUs will
just round it to zero (!)

Thanks to @alexneufeld for tracking this down in
https://github.com/mkeeter/halfspace/issues/1

+++ fidget-raster/src/lib.rs
@@ -420,7 +420,10 @@ impl<P, S: ImageSizeLike> std::ops::IndexMut<(usize, usize)> for Image<P, S> {
 #[derive(Debug, Default, Copy, Clone, IntoBytes, FromBytes, Immutable)]
 pub struct GeometryPixel {
     /// Z position of this pixel, in voxel units
-    pub depth: u32, // TODO should this be `f32`?
+    ///
+    /// The fractional component is always zero. Empty pixels always have a
+    /// depth of 0.
+    pub depth: f32,
     /// Function gradients at this pixel
     pub normal: [f32; 3],
 }
